refactor: move LineStyle and BorderSymbol to merge, change visibility to pub(crate)
