+++ ratatui-core/src/symbols/line.rs
@@ -1,5 +1,3 @@
-use core::str::FromStr;
-
 pub const VERTICAL: &str = "│";
 pub const DOUBLE_VERTICAL: &str = "║";
 pub const THICK_VERTICAL: &str = "┃";
@@ -59,299 +57,6 @@ pub const THICK_HORIZONTAL_UP: &str = "┻";
 pub const CROSS: &str = "┼";
 pub const DOUBLE_CROSS: &str = "╬";
 pub const THICK_CROSS: &str = "╋";
-
-/// A visual style defining the appearance of a single line making up a block border.
-#[derive(PartialEq, Debug, Clone, Copy)]
-pub enum LineStyle {
-    /// Represents the absence of a line.
-    Nothing,
-
-    /// A single line (e.g. `─`, `│`).
-    Plain,
-
-    /// A rounded line style, only applicable in corner symbols (e.g. `╭`, `╯`).
-    Rounded,
-
-    /// A double line (e.g. `═`, `║`).
-    Double,
-
-    /// A thickened line (e.g. `━`, `┃`).
-    Thick,
-
-    /// A dashed line with a double dash pattern (e.g. `╌`, `╎`).
-    DoubleDash,
-
-    /// A dashed line with a triple dash pattern (e.g. `┄`, `┆`).
-    TripleDash,
-
-    /// A thicker variant of the triple dash (e.g. `┅`, `┇`).
-    TripleDashThick,
-
-    /// A dashed line with four dashes (e.g. `┈`, `┊`).
-    QuadrupleDash,
-
-    /// A thicker variant of the quadruple dash (e.g. `┉`, `┋`).
-    QuadrupleDashThick,
-}
-
-impl LineStyle {
-    /// Merges line styles.
-    #[must_use]
-    pub fn merge(self, other: Self) -> Self {
-        if other == Self::Nothing { self } else { other }
-    }
-}
-
-/// Represents a composite border symbol using individual line components.
-#[derive(Debug, Copy, Clone)]
-pub struct BorderSymbol {
-    pub right: LineStyle,
-    pub up: LineStyle,
-    pub left: LineStyle,
-    pub down: LineStyle,
-}
-
-impl BorderSymbol {
-    /// Creates a new `BorderSymbol`, based on individual line styles.
-    #[must_use]
-    pub const fn new(right: LineStyle, up: LineStyle, left: LineStyle, down: LineStyle) -> Self {
-        Self {
-            right,
-            up,
-            left,
-            down,
-        }
-    }
-    /// Finds the closest representation of the `BorderSymbol`, that has a corresponding unicode
-    /// character.
-    #[must_use]
-    pub fn best_fit(mut self) -> Self {
-        use LineStyle::{Double, Plain, Rounded, Thick};
-        // Check if we got a character that can be displayed after change
-        if TryInto::<&str>::try_into(self).is_ok() {
-            return self;
-        }
-        // There is no character that combines double and thick
-        // There is more characters for thick borders.
-        if self.contains(Double) && self.contains(Thick) {
-            self = self.replace(Double, Thick);
-        }
-        if TryInto::<&str>::try_into(self).is_ok() {
-            return self;
-        }
-        if self.contains(Double) {
-            self = self.replace(Double, Plain);
-        }
-        if TryInto::<&str>::try_into(self).is_ok() {
-            return self;
-        }
-        // Rouned border character are only available for corners.
-        if self.contains(Rounded) {
-            self = self.replace(Rounded, Plain);
-        }
-
-        self
-    }
-
-    /// Checks if any of the line components making the `BorderSymbol` matches the `style`.
-    pub fn contains(self, style: LineStyle) -> bool {
-        self.up == style || self.right == style || self.down == style || self.left == style
-    }
-
-    /// Replaces all line styles matching `from` by `to`.
-    #[must_use]
-    pub fn replace(mut self, from: LineStyle, to: LineStyle) -> Self {
-        self.up = if self.up == from { to } else { self.up };
-        self.right = if self.right == from { to } else { self.right };
-        self.down = if self.down == from { to } else { self.down };
-        self.left = if self.left == from { to } else { self.left };
-        self
-    }
-}
-
-impl FromStr for BorderSymbol {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        s.try_into()
-    }
-}
-
-// Defines a translation between `BorderSymbol` and the corresponding character.
-macro_rules! define_symbols {
-    (
-        $( $symbol:expr => ($right:ident, $up:ident, $left:ident, $down:ident) ),* $(,)?
-    ) => {
-
-        impl TryFrom<&str> for BorderSymbol {
-            type Error = ();
-            fn try_from(value: &str) -> Result<Self, Self::Error> {
-                use LineStyle::*;
-                match value {
-                    $( $symbol => Ok(Self::new($right, $up, $left, $down)) ),* ,
-                    _ => Err(()),
-                }
-            }
-        }
-
-        impl TryFrom<BorderSymbol> for &'static str {
-            type Error = ();
-            fn try_from(value: BorderSymbol) -> Result<Self, Self::Error> {
-                use LineStyle::*;
-                match (value.right, value.up, value.left, value.down) {
-                    $( ($right, $up, $left, $down) => Ok($symbol) ),* ,
-                    _ => Err(()),
-                }
-            }
-        }
-    };
-}
-
-define_symbols!(
-    "─" => (Plain, Nothing, Plain, Nothing),
-    "━" => (Thick, Nothing, Thick, Nothing),
-    "│" => (Nothing, Plain, Nothing, Plain),
-    "┃" => (Nothing, Thick, Nothing, Thick),
-
-    "┄" => (TripleDash, Nothing, TripleDash, Nothing),
-    "┅" => (TripleDashThick, Nothing, TripleDashThick, Nothing),
-    "┆" => (Nothing, TripleDash, Nothing, TripleDash),
-    "┇" => (Nothing, TripleDashThick, Nothing, TripleDashThick),
-    "┈" => (QuadrupleDash, Nothing, QuadrupleDash, Nothing),
-    "┉" => (QuadrupleDashThick, Nothing, QuadrupleDashThick, Nothing),
-    "┊" => (Nothing, QuadrupleDash, Nothing, QuadrupleDash),
-    "┋" => (Nothing, QuadrupleDashThick, Nothing, QuadrupleDashThick),
-
-    "┘" => (Nothing, Plain, Plain, Nothing),
-    "┚" => (Nothing, Thick, Plain, Nothing),
-    "┙" => (Nothing, Plain, Thick, Nothing),
-    "┛" => (Nothing, Thick, Thick, Nothing),
-
-    "┐" => (Nothing, Nothing, Plain, Plain),
-    "┒" => (Nothing, Nothing, Plain, Thick),
-    "┑" => (Nothing, Nothing, Thick, Plain),
-    "┓" => (Nothing, Nothing, Thick, Thick),
-
-    "┌" => (Plain, Nothing, Nothing, Plain),
-    "┍" => (Thick, Nothing, Nothing, Plain),
-    "┎" => (Plain, Nothing, Nothing, Thick),
-    "┏" => (Thick, Nothing, Nothing, Thick),
-
-    "└" => (Plain, Plain, Nothing, Nothing),
-    "┕" => (Thick, Plain, Nothing, Nothing),
-    "┖" => (Plain, Thick, Nothing, Nothing),
-    "┗" => (Thick, Thick, Nothing, Nothing),
-
-    "┴" => (Plain, Plain, Plain, Nothing),
-    "┸" => (Plain, Thick, Plain, Nothing),
-    "┵" => (Plain, Plain, Thick, Nothing),
-    "┶" => (Thick, Plain, Plain, Nothing),
-    "┺" => (Thick, Thick, Plain, Nothing),
-    "┹" => (Plain, Thick, Thick, Nothing),
-    "┷" => (Thick, Plain, Thick, Nothing),
-
-    "┬" => (Plain, Nothing, Plain, Plain),
-    "┮" => (Thick, Nothing, Plain, Plain),
-    "┰" => (Plain, Nothing, Plain, Thick),
-    "┭" => (Plain, Nothing, Thick, Plain),
-    "┱" => (Plain, Nothing, Thick, Thick),
-    "┲" => (Thick, Nothing, Plain, Thick),
-    "┯" => (Thick, Nothing, Thick, Plain),
-    "┳" => (Thick, Nothing, Thick, Thick),
-
-    "┤" => (Nothing, Plain, Plain, Plain),
-    "┦" => (Nothing, Thick, Plain, Plain),
-    "┥" => (Nothing, Plain, Thick, Plain),
-    "┧" => (Nothing, Plain, Plain, Thick),
-    "┨" => (Nothing, Thick, Plain, Thick),
-    "┪" => (Nothing, Plain, Thick, Thick),
-    "┩" => (Nothing, Thick, Thick, Plain),
-    "┫" => (Nothing, Thick, Thick, Thick),
-
-    "├" => (Plain, Plain, Nothing, Plain),
-    "┞" => (Plain, Thick, Nothing, Plain),
-    "┝" => (Thick, Plain, Nothing, Plain),
-    "┟" => (Plain, Plain, Nothing, Thick),
-    "┡" => (Thick, Thick, Nothing, Plain),
-    "┢" => (Thick, Plain, Nothing, Thick),
-    "┠" => (Plain, Thick, Nothing, Thick),
-    "┣" => (Thick, Thick, Nothing, Thick),
-
-    "┽" => (Plain, Plain, Thick, Plain),
-    "┼" => (Plain, Plain, Plain, Plain),
-    "╁" => (Plain, Plain, Plain, Thick),
-    "╅" => (Plain, Plain, Thick, Thick),
-    "╃" => (Plain, Thick, Thick, Plain),
-    "╀" => (Plain, Thick, Plain, Plain),
-    "╂" => (Plain, Thick, Plain, Thick),
-    "┾" => (Thick, Plain, Plain, Plain),
-    "╆" => (Thick, Plain, Plain, Thick),
-    "┿" => (Thick, Plain, Thick, Plain),
-    "╄" => (Thick, Thick, Plain, Plain),
-    "╈" => (Thick, Plain, Thick, Thick),
-    "╊" => (Thick, Thick, Plain, Thick),
-    "╉" => (Plain, Thick, Thick, Thick),
-    "┻" => (Thick, Thick, Thick, Nothing),
-    "╇" => (Thick, Thick, Thick, Plain),
-    "╋" => (Thick, Thick, Thick, Thick),
-
-    "═" => (Double, Nothing, Double, Nothing),
-    "║" => (Nothing, Double, Nothing, Double),
-
-    "╓" => (Plain, Nothing, Nothing, Double),
-    "╒" => (Double, Nothing, Nothing, Plain),
-    "╔" => (Double, Nothing, Nothing, Double),
-
-    "╕" => (Nothing, Nothing, Double, Plain),
-    "╖" => (Nothing, Nothing, Plain, Double),
-    "╗" => (Nothing, Nothing, Double, Double),
-
-    "╘" => (Double, Plain, Nothing, Nothing),
-    "╙" => (Plain, Double, Nothing, Nothing),
-    "╚" => (Double, Double, Nothing, Nothing),
-
-    "╛" => (Nothing, Plain, Double, Nothing),
-    "╜" => (Nothing, Double, Plain, Nothing),
-    "╝" => (Nothing, Double, Double, Nothing),
-
-    "╞" => (Double, Plain, Nothing, Plain),
-    "╟" => (Plain, Double, Nothing, Double),
-    "╠" => (Double, Double, Nothing, Double),
-
-    "╡" => (Nothing, Plain, Double, Plain),
-    "╢" => (Nothing, Double, Plain, Double),
-    "╣" => (Nothing, Double, Double, Double),
-
-    "╤" => (Double, Nothing, Double, Plain),
-    "╥" => (Plain, Nothing, Plain, Double),
-    "╦" => (Double, Nothing, Double, Double),
-
-    "╧" => (Double, Plain, Double, Nothing),
-    "╨" => (Plain, Double, Plain, Nothing),
-    "╩" => (Double, Double, Double, Nothing),
-
-    "╪" => (Double, Plain, Double, Plain),
-    "╫" => (Plain, Double, Plain, Double),
-    "╬" => (Double, Double, Double, Double),
-
-    "╭" => (Rounded, Nothing, Nothing, Rounded),
-    "╮" => (Nothing, Nothing, Rounded, Rounded),
-    "╯" => (Nothing, Rounded, Rounded, Nothing),
-    "╰" => (Rounded, Rounded, Nothing, Nothing),
-
-    "╴" => (Nothing, Nothing, Plain, Nothing),
-    "╵" => (Nothing, Plain, Nothing, Nothing),
-    "╶" => (Plain, Nothing, Nothing, Nothing),
-    "╷" => (Nothing, Nothing, Nothing, Plain),
-    "╸" => (Nothing, Nothing, Thick, Nothing),
-    "╹" => (Nothing, Thick, Nothing, Nothing),
-    "╺" => (Thick, Nothing, Nothing, Nothing),
-    "╻" => (Nothing, Nothing, Nothing, Thick),
-    "╼" => (Thick, Nothing, Plain, Nothing),
-    "╽" => (Nothing, Plain, Nothing, Thick),
-    "╾" => (Plain, Nothing, Thick, Nothing),
-    "╿" => (Nothing, Thick, Nothing, Plain),
-);
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
 pub struct Set {
